refactor: simplify project structure and remove uneeded code.
