--- core/src/main/java/io/ap4k/project/GenericInfo.java
@@ -1,41 +1,0 @@
-/**
- * Copyright (C) 2018 Ioannis Canellos 
- *     
- * 
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *         http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * 
-**/
-package io.ap4k.project;
-
-import java.nio.file.Path;
-
-public class GenericInfo implements BuildInfo {
-
-    private static final String LATEST = "latest";
-    private final Path root;
-
-    public GenericInfo(Path root) {
-        this.root = root;
-    }
-
-    @Override
-    public String getName() {
-        return root.getFileName().toString();
-    }
-
-    @Override
-    public String getVersion() {
-        return LATEST;
-    }
-
-}
