--- core/src/main/java/io/ap4k/project/SbtBuildInfo.java
@@ -1,41 +1,0 @@
-/**
- * Copyright (C) 2018 Ioannis Canellos 
- *     
- * 
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * 
- *         http://www.apache.org/licenses/LICENSE-2.0
- * 
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- * 
-**/
-package io.ap4k.project;
-
-public class SbtBuildInfo implements BuildInfo {
-
-    @Override
-    public String getName() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public String getVersion() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public String getPackaging() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public String getOutputFileName() {
-        throw new UnsupportedOperationException();
-    }
-}
